Allow to define default instances for cluster configuration classes
Cluster configuration classes are simple Jackson-annotated POJOs which are being handled by
ClusterConfigService. This changeset introduces the possibility to define a default instance
being used if there is no explicit instance in the database.

The default instance must be provided by a static method in the respective cluster configuration
POJO with the following signature (with `T` being the specific cluster configuration class):

```
public static T defaultConfig();
```